feat: add support for showing overlay with keyboard nav
