+++ styles/globals.css
@@ -2,7 +2,16 @@
   accent-color: auto;
 }
 
+html,
+body,
+body > div:first-child,
+div#__next,
+div#__next > div {
+  height: 100%;
+}
+
 ul {
-  margin-block: 0;
-  padding-inline-start: 0;
+  padding: 0;
+  margin: 0;
+  list-style-type: none;
 }
